Create hiccup module
