+++ gizmos/hiccup.py
@@ -1,0 +1,69 @@
+def curie2href(curie):
+    """Convert a CURIE to an HREF"""
+    return f"?id={curie}".replace("#", "%23")
+
+
+def render(prefixes, element, depth=0):
+    """Render hiccup-style HTML vector as HTML."""
+    indent = "  " * depth
+    if not isinstance(element, list):
+        raise Exception(f"Element is not a list: {element}")
+    if len(element) == 0:
+        raise Exception(f"Element is an empty list")
+    tag = element.pop(0)
+    if not isinstance(tag, str):
+        raise Exception(f"Tag '{tag}' is not a string in '{element}'")
+    output = f"{indent}<{tag}"
+
+    if len(element) > 0 and isinstance(element[0], dict):
+        attrs = element.pop(0)
+        if tag == "a" and "href" not in attrs and "resource" in attrs:
+            attrs["href"] = curie2href(attrs["resource"])
+        for key, value in attrs.items():
+            if key in ["checked"]:
+                if value:
+                    output += f" {key}"
+            else:
+                output += f' {key}="{value}"'
+
+    if tag in ["meta", "link"]:
+        output += "/>"
+        return output
+    output += ">"
+    spacing = ""
+    if len(element) > 0:
+        for child in element:
+            if isinstance(child, str):
+                output += child
+            elif isinstance(child, list):
+                try:
+                    output += "\n" + render(prefixes, child, depth=depth + 1)
+                    spacing = f"\n{indent}"
+                except Exception as e:
+                    raise Exception(f"Bad child in '{element}'", e)
+            else:
+                raise Exception(f"Bad type for child '{child}' in '{element}'")
+    output += f"{spacing}</{tag}>"
+    return output
+
+
+def render_text(element):
+    """Render hiccup-style HTML vector as text."""
+    if not isinstance(element, list):
+        raise Exception(f"Element is not a list: {element}")
+    if len(element) == 0:
+        raise Exception(f"Element is an empty list")
+    tag = element.pop(0)
+    output = ""
+    if len(element) > 0:
+        for child in element:
+            if isinstance(child, str):
+                output += child
+            elif isinstance(child, list):
+                try:
+                    output += render_text(child)
+                except Exception as e:
+                    raise Exception(f"Bad child in '{element}'", e)
+            else:
+                raise Exception(f"Bad type for child '{child}' in '{element}'")
+    return output
